Add files via upload
This project provides the facility of packing and unpacking the regular files.                                                      
In case of packing activity all the data of multiple regular files gets stored in single file with all related metadata. In case of unpacking activity the data from the packed file gets extracted in the separate file with all necessary details. 
To maintain the security of the data we provide encryption and decryption techniques.                                             
We provide Graphical user interface to interact with the application